requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-safedelete==1.1.2
djangorestframework==3.13.1
docopt==0.6.2
num2words==0.5.10
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
